requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
supabase==2.3.0
python-dotenv==1.0.0
asyncio==3.4.3
requests==2.31.0
lxml==4.9.3
playwright==1.41.2 